views.py files of all apps are changed. Added simple authorizations for Update and Delete methods mostly

+++ backend/service_requests/views.py
@@ -6,7 +6,7 @@ from rest_framework.views import APIView
 
 from .models import Request
 from .serializers import RequestSerializer
-from users.models import User, RegularUser
+from users.models import RegularUser
 
 
 
@@ -32,22 +32,28 @@ class RequestDetail(APIView):
             raise Http404
 
     def get(self, request, pk, format=None):
-        r = self.get_object(pk)
+        r = self.get_object(pk) # request object with id = pk
         serializer = RequestSerializer(r)
         return Response(serializer.data)
 
     def put(self, request, pk, format=None):
         r = self.get_object(pk)
-        serializer = RequestSerializer(r, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_staff or request.user.id == r.user.user.id:            
+            serializer = RequestSerializer(r, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
     def delete(self, request, pk, format=None):
-        r = self.get_object(pk)
-        r.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        r = self.get_object(pk) # request object with id = pk
+        
+        if request.user.is_staff or request.user.id == r.user.user.id:
+            r.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(["GET", "POST"])
@@ -64,19 +70,38 @@ def all_requests(request):
         return Response(serializer.data)
 
 
-@api_view(["GET", "PUT", ""])
+@api_view(["GET", "PUT", "DELETE"])
 def request_detail(request, pk):
-    r = Request.objects.get(id = pk)
-    serializer = RequestSerializer(r, many = False)
+    r = Request.objects.get(id = pk) # request object with id = pk
 
-    return Response(serializer.data)
+    if request.method == "GET":
+        serializer = RequestSerializer(r, many = False)
+        return Response(serializer.data)
+    
+    elif request.method == "DELETE":
+        if request.user.is_staff or request.user.id == r.user.user.id:
+            r.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+    
+    elif request.method == "PUT":
+        if request.user.is_staff or request.user.id == r.user.user.id:            
+            serializer = RequestSerializer(r, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(["GET"])
 def user_requests(request, username):
-    general_user = User.objects.get(username = username)
-    user = RegularUser.objects.get(user = general_user)
-    requests = Request.objects.filter(user = user)
-    serializer = RequestSerializer(requests, many = True)
+    if request.user.username == username or request.user.is_staff:
+        user = RegularUser.objects.get(user__username = username)
+        requests = Request.objects.filter(user = user)
+        serializer = RequestSerializer(requests, many = True)
 
-    return Response(serializer.data)
+        return Response(serializer.data)
+
+    return Response(status=status.HTTP_401_UNAUTHORIZED)
